create a shared package (for code shared between frontend and backend)
It's different from "common" in that, "common" is for components shared
between different frontend apps and "shared" is mostly for constants and
things that need to be shared between frontend and backend. It also uses
commonjs bundles and isn't covered by cra's build process.

This might need a rework or at least better naming but it'll do for now